fix: linkify URLs containing scheme-like words
